requirements.txt
streamlit
firebase-admin
pytesseract==0.3.10
Pillow==10.0.1
pandas==2.1.3
opencv-python==4.8.1.78
numpy==1.24.3
python-dateutil==2.8.2
tqdm==4.66.1
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
# Pinned stable version for video processing
moviepy==1.0.3 
SpeechRecognition
psutil
reportlab
matplotlib
seaborn

# New dependencies for Advanced AI and Legal Analysis
google-genai 
scikit-learn
langchain